requirements.txt
asgiref==3.5.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.0.6
django-bootstrap-pagination==1.7.1
django-cors-headers==4.0.0
django-debug-toolbar==4.0.0
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.5
fonttools==4.39.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mypy-extensions==1.0.0
numpy==1.24.2
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
transliterate==1.10.2
uritemplate==4.1.1
urllib3==1.26.15
